Moves instance code from utility to instance files so utility files just about utilities
